task: Reorganize folders
- rename common path to application.
- register HealthController and AppConfigModule directly on AppModule.
- drop CommonModule.
- auto-load configs in AppConfigModule.
