+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "nest-base",
-  "version": "1.8.22",
+  "version": "1.9.0",
   "description": "",
   "license": "MIT",
   "keywords": [],
