+++ src/application/constants/index.ts
