+++ src/application/dto/get-health-response.dto.ts
